implement logic to conditionally render icons
